requirements.txt
overloading
mysql-connector-python
aenum
easydb
PyQt5
PyQt5-sip
